refactor: GameController -> GameRunner 이름 변경
