+++ src/main/java/blackjack/Application.java
@@ -1,9 +1,7 @@
 package blackjack;
 
-import blackjack.controller.GameController;
-
 public class Application {
     public static void main(String[] args) {
-        new GameController().run();
+        new GameRunner().run();
     }
 }
